+++ src/screens/setting/favoriteList/EditFavoriteScreen.tsx
