+++ src/screens/setting/favoriteList/FavoriteListScreen.tsx
