+++ src/screens/setting/nameTag/ProfileNameTag.tsx
@@ -9,8 +9,11 @@ import {
 import {ProfileUpdateIcon} from '~/assets/images';
 import {TouchableOpacity} from 'react-native';
 import {useUserInfo} from '~/zustand/auth/auth';
+import {useNavigation} from '@react-navigation/native';
+import {RootStackNavigationProp} from '~/App';
 
 const ProfileNameTag = () => {
+  const navigation = useNavigation<RootStackNavigationProp>();
   const tag = `<svg width="400" height="74" viewBox="0 0 400 74" fill="none" xmlns="http://www.w3.org/2000/svg">
   <path d="M376.451 37.1406L399.096 73.25H0.5V0.5H399.096L376.451 36.6094L376.285 36.875L376.451 37.1406Z" fill="white" stroke="#FF6F61"/>
   <rect width="11.25" height="73.75" fill="#FF6F61"/>
@@ -26,7 +29,7 @@ const ProfileNameTag = () => {
       <WordContainer>
         <Wrapper>
           <ProfileWrapper>
-            <Propfile
+            <Profile
               source={{uri: `data:image/png;base64,${userInfo.profile}`}}
               resizeMode="contain"
               alt={'이미지 읽기 실패'}
@@ -41,7 +44,7 @@ const ProfileNameTag = () => {
             </UserInfoRow>
             <Email>{userInfo.email}</Email>
           </UserInfoColumn>
-          <TouchableOpacity>
+          <TouchableOpacity onPress={() => navigation.navigate('EditProfile')}>
             <ProfileUpdateIcon />
           </TouchableOpacity>
         </Wrapper>
@@ -116,7 +119,7 @@ const ProfileWrapper = styled.View`
   overflow: hidden;
 `;
 
-const Propfile = styled.Image`
+const Profile = styled.Image`
   width: 100%;
   height: 100%;
   align-items: center;
